+++ copy_with_extension_gen/test/generated_code_test_cases/test_cases_exceptions.dart
@@ -24,3 +24,17 @@ class WrongConstructor {}
 class NoDefaultConstructor {
   NoDefaultConstructor.nonDefault();
 }
+
+@ShouldThrow(
+  'The nullability of the constructor parameter "nullableWithNonNullableConstructor" does not match the nullability of the corresponding field in the object.',
+)
+@CopyWith()
+class TestNullability {
+  TestNullability(
+    int this.nullableWithNonNullableConstructor,
+  );
+
+  // Some info on this: https://github.com/numen31337/copy_with_extension/pull/69
+  // If a field is nullable, you can change the type of the constructor parameter to be non-nullable. However, if you do this, an exception may be thrown because the constructor only accepts non-nullable parameters. The issue is that we cannot guarantee that the parameter will be non-null when calling the constructor in the `copyWith` function. Therefore, even if the constructor constructs an object with a nullable field, it is currently impossible to achieve this in the implementation.
+  final int? nullableWithNonNullableConstructor;
+}